--- packages/grida-tokens/index.ts
@@ -1,1 +1,0 @@
-export * from "./tokens";
